+++ attributes/default.rb
@@ -1,3 +1,4 @@
 default['sanitize']['iptables'] = true
 default['sanitize']['install_packages'] = []
 default['sanitize']['apt_repositories'] = {}
+default['sanitize']['ports']['ssh'] = true